ecommerce-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
email-validator==2.1.0
reportlab==4.0.7
python-multipart==0.0.6
bcrypt==4.1.2
pyjwt==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.36
alembic==1.12.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1